Creación del componente LazyImage
Se creó el componente lazyImage reutilizable. En cuanto al Detail, no es mi jurisdicción trabajar en los estilos, puesto que es parte de otra tarea ajena a mí.
